refactor(windows): split recovery policy helpers
