+++ SSRVPN_Windows/windows/runner/launcher_main.cpp
@@ -7,6 +7,7 @@
 #include <string>
 #include <vector>
 
+#include "guardian_restart.h"
 #include "system_proxy_recovery.h"
 
 namespace {
@@ -1046,27 +1047,8 @@ bool StartGuardian(const std::wstring& launcher_path,
   }
   return true;
 }
-
 constexpr DWORD kGuardianRestartAttempts = 5;
 constexpr DWORD kGuardianRestartDelayMs = 100;
-
-bool StartReplacementGuardianWithRetry(
-    const std::wstring& launcher_path,
-    const std::wstring& launcher_directory, DWORD child_process_id,
-    HANDLE* guardian_process, HANDLE* guardian_commit_event,
-    DWORD* error_code) {
-  for (DWORD attempt = 0; attempt < kGuardianRestartAttempts; ++attempt) {
-    if (StartGuardian(launcher_path, launcher_directory, child_process_id, 0,
-                      guardian_process, guardian_commit_event, error_code)) {
-      return true;
-    }
-    if (attempt + 1 < kGuardianRestartAttempts) {
-      ::Sleep(kGuardianRestartDelayMs);
-    }
-  }
-  return false;
-}
-
 // ── Child process creation ──
 
 bool CreateChildProcess(const std::wstring& child_path,
@@ -1406,9 +1388,12 @@ int APIENTRY wWinMain(_In_ HINSTANCE instance, _In_opt_ HINSTANCE previous,
     if (!fail_closed_cleanup_pending) {
       HANDLE replacement_commit_event = nullptr;
       if (process_job != nullptr &&
-          StartReplacementGuardianWithRetry(
-              launcher_path, launcher_directory, child_process_id,
-              &guardian_process, &replacement_commit_event, &error)) {
+          RetryGuardianStart(kGuardianRestartAttempts,
+              kGuardianRestartDelayMs, [&]() {
+                return StartGuardian(
+                    launcher_path, launcher_directory, child_process_id, 0,
+                    &guardian_process, &replacement_commit_event, &error);
+              })) {
         if (::SetEvent(replacement_commit_event)) {
           ::CloseHandle(replacement_commit_event);
           continue;
